requirements.txt
psycopg2-binary
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
attrs==25.3.0
beautifulsoup4==4.8.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
compressed-rtf==1.0.7
dataclasses-json==0.6.7
diskcache==5.6.3
docx2txt==0.9
ebcdic==1.1.1
extract-msg==0.28.7
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8
langsmith==0.3.31
llama_cpp_python==0.3.8
lxml==5.3.2
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.2.4
olefile==0.47
orjson==3.10.16
packaging==24.2
pdfminer.six==20191110
pillow==11.2.1
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
python-pptx==0.6.23
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.40
starlette==0.46.2
striprtf==0.0.29
tenacity==9.1.2
tika==3.1.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.1
xlrd==1.2.0
XlsxWriter==3.2.2
yarl==1.19.0
zstandard==0.23.0
fastapi_admin
langchain_openai
chromadb
# psycopg2
sentence-transformers
tiktoken
sentencepiece
faiss-cpu
alembic
sqladmin
pydantic[email]
six==1.16.0  
striprtf
langchain-huggingface
passlib
PyJWT
pytest